Split atomicpuppy.py into subscriptions and streams
