+++ tests/test_stream_config.py
@@ -1,6 +1,6 @@
 import io
 
-from atomicpuppy.atomicpuppy import make_subscription_config
+from atomicpuppy.subscriptions import make_subscription_config
 from atomicpuppy.events import EventCounter, RedisCounter
 import platform
